refactor: begin routes-ui.js extraction into ui-routes/ modules (A2)
Decompose the lib/routes-ui.js god file (tech-debt item A2). This first
milestone adds guardrails and extracts the smallest, most isolated route
groups. All route/helper moves are behavior-preserving (verbatim).

Guardrails (run on every future batch):
- test/ui-routes-table-test.js: in-process snapshot of the exact (method, path)
  set registered by routes-ui.js (128 routes) - the invariant proving no route
  is dropped, added, or duplicated by a move.
- test/ui-routes-smoke-test.js: HTTP reachability of every parameterless
  /admin/* GET (not 404, not 5xx).
- test/ui-routes-size-test.js: line-count ratchet so routes-ui.js cannot regrow.

Extractions:
- lib/ui-routes/route-helpers.js: shared pure helpers used across route groups
  (throwAsBoom, formatAccountData, formatServerState, getServerStatus,
  getOpenAiModels/OPEN_AI_MODELS, getExampleDocumentsPayloads, cachedTemplates),
  imported back into routes-ui.js.
- lib/ui-routes/smtp-test-routes.js: /admin/smtp/create-test, /admin/smtp/check-test.
- lib/ui-routes/unsubscribe-routes.js: public /unsubscribe, /unsubscribe/address.

routes-ui.js: 8838 -> 8143 lines. Wiring mirrors the existing
admin-entities-routes.js pattern. Route table verified unchanged via the new
snapshot test, full server boot, and the api/ui smoke tests.
