+++ lib/routes-ui.js
@@ -35,7 +35,6 @@ const pbkdf2 = require('@phc/pbkdf2');
 const { Account } = require('./account');
 const { Gateway } = require('./gateway');
 const { redis, submitQueue, notifyQueue, documentsQueue } = require('./db');
-const psl = require('psl');
 const { oauth2Apps, OAUTH_PROVIDERS, oauth2ProviderData, SERVICE_ACCOUNT_PROVIDERS } = require('./oauth2-apps');
 const { verifyOAuth2App } = require('./oauth/verify-app');
 const { autodetectImapSettings } = require('./autodetect-imap-settings');
@@ -51,13 +50,11 @@ const {
     exportIdSchema
 } = require('./schemas');
 const fs = require('fs');
-const pathlib = require('path');
 const timezonesList = require('timezones-list').default;
 const { Client: ElasticSearch } = require('@elastic/elasticsearch');
 const { llmPreProcess } = require('./llm-pre-process');
 const { locales } = require('./translations');
 const capa = require('./capa');
-const exampleDocumentsPayloads = require('./payload-examples-documents.json');
 const { defaultMappings } = require('./es');
 const { getESClient } = require('../lib/document-store');
 const assert = require('assert');
@@ -68,6 +65,17 @@ const { simpleParser } = require('mailparser');
 const libmime = require('libmime');
 
 const adminEntitiesRoutes = require('./ui-routes/admin-entities-routes');
+const smtpTestRoutes = require('./ui-routes/smtp-test-routes');
+const unsubscribeRoutes = require('./ui-routes/unsubscribe-routes');
+const {
+    throwAsBoom,
+    formatAccountData,
+    getServerStatus,
+    getOpenAiModels,
+    OPEN_AI_MODELS,
+    getExampleDocumentsPayloads,
+    cachedTemplates
+} = require('./ui-routes/route-helpers');
 const { Export } = require('./export');
 const passkeys = require('./passkeys');
 const {
@@ -115,24 +123,6 @@ const DISABLE_MESSAGE_BROWSER = getBoolean(readEnvValue('EENGINE_DISABLE_MESSAGE
 const { fetch: fetchCmd } = require('undici');
 
 const LICENSE_HOST = 'https://postalsys.com';
-const SMTP_TEST_HOST = 'https://api.nodemailer.com';
-
-const OPEN_AI_MODELS = [
-    {
-        name: 'GPT-3 (instruct)',
-        id: 'gpt-3.5-turbo-instruct'
-    },
-
-    {
-        name: 'GPT-3 (chat)',
-        id: 'gpt-3.5-turbo'
-    },
-
-    {
-        name: 'GPT-4',
-        id: 'gpt-4'
-    }
-];
 
 const IMAP_INDEXERS = [
     {
@@ -184,27 +174,6 @@ const notificationTypes = Object.keys(consts)
         name: consts[`${key}_NOTIFY`],
         description: consts[`${key}_DESCRIPTION`]
     }));
-
-const cachedTemplates = {
-    addressList: fs.readFileSync(pathlib.join(__dirname, '..', 'views', 'partials', 'address_list.hbs'), 'utf-8'),
-    testSend: fs.readFileSync(pathlib.join(__dirname, '..', 'views', 'partials', 'test_send.hbs'), 'utf-8')
-};
-
-const getOpenAiModels = async (models, selectedModel) => {
-    let modelList = (await settings.get('openAiModels')) || structuredClone(models);
-
-    if (selectedModel && !modelList.find(model => model.id === selectedModel)) {
-        modelList.unshift({
-            name: selectedModel,
-            id: selectedModel
-        });
-    }
-
-    return modelList.map(model => {
-        model.selected = model.id === selectedModel;
-        return model;
-    });
-};
 
 const FIELD_TYPES = [
     {
@@ -363,206 +332,6 @@ const OKTA_OAUTH2_CLIENT_ID = readEnvValue('OKTA_OAUTH2_CLIENT_ID');
 const OKTA_OAUTH2_CLIENT_SECRET = readEnvValue('OKTA_OAUTH2_CLIENT_SECRET');
 const USE_OKTA_AUTH = !!(OKTA_OAUTH2_ISSUER && OKTA_OAUTH2_CLIENT_ID && OKTA_OAUTH2_CLIENT_SECRET);
 
-function formatAccountData(account, gt) {
-    account.type = {};
-
-    if (account.oauth2 && account.oauth2.app) {
-        let providerData = oauth2ProviderData(account.oauth2.app.provider);
-        account.type = providerData;
-    } else if (account.oauth2 && account.oauth2.provider) {
-        account.type = oauth2ProviderData(account.oauth2.provider);
-    } else if (account.imap && !account.imap.disabled) {
-        account.type.icon = 'fa fa-envelope-square';
-        account.type.name = 'IMAP';
-        account.type.comment = psl.get(account.imap.host) || account.imap.host;
-    } else if (account.smtp) {
-        account.type.icon = 'fa fa-paper-plane';
-        account.type.name = 'SMTP';
-        account.type.comment = psl.get(account.smtp.host) || account.smtp.host;
-    } else if (account.oauth2 && account.oauth2.auth && account.oauth2.auth.delegatedAccount) {
-        account.type.icon = 'fa fa-arrow-alt-circle-right';
-        account.type.name = gt.gettext('Delegated');
-        account.type.comment = util.format(gt.gettext('Using credentials from "%s"'), account.oauth2.auth.delegatedAccount);
-    } else {
-        account.type.name = 'N/A';
-    }
-
-    switch (account.state) {
-        case 'init':
-            account.stateLabel = {
-                type: 'info',
-                name: 'Initializing',
-                spinner: true
-            };
-            break;
-
-        case 'connecting':
-            account.stateLabel = {
-                type: 'info',
-                name: 'Connecting'
-            };
-            break;
-
-        case 'syncing':
-            account.stateLabel = {
-                type: 'info',
-                name: 'Syncing',
-                spinner: true
-            };
-            break;
-
-        case 'connected':
-            account.stateLabel = {
-                type: 'success',
-                name: 'Connected'
-            };
-            break;
-
-        case 'disabled':
-            account.stateLabel = {
-                type: 'secondary',
-                name: 'Disabled',
-                error: account.disabledReason
-            };
-            break;
-
-        case 'authenticationError':
-        case 'connectError': {
-            let errorMessage = account.lastErrorState ? account.lastErrorState.response : false;
-            if (account.lastErrorState) {
-                switch (account.lastErrorState.serverResponseCode) {
-                    case 'ETIMEDOUT':
-                        errorMessage = gt.gettext('Connection timed out. This usually occurs if you are behind a firewall or connecting to the wrong port.');
-                        break;
-                    case 'ClosedAfterConnectTLS':
-                        errorMessage = gt.gettext('The server unexpectedly closed the connection.');
-                        break;
-                    case 'ClosedAfterConnectText':
-                        errorMessage = gt.gettext(
-                            'The server unexpectedly closed the connection. This usually happens when attempting to connect to a TLS port without TLS enabled.'
-                        );
-                        break;
-                    case 'ECONNREFUSED':
-                        errorMessage = gt.gettext(
-                            'The server refused the connection. This typically occurs if the server is not running, is overloaded, or you are connecting to the wrong host or port.'
-                        );
-                        break;
-                }
-            }
-
-            account.stateLabel = {
-                type: 'danger',
-                name: 'Failed',
-                error: errorMessage
-            };
-            break;
-        }
-        case 'unset':
-            account.stateLabel = {
-                type: 'light',
-                name: 'Not syncing'
-            };
-            break;
-        case 'disconnected':
-            account.stateLabel = {
-                type: 'warning',
-                name: 'Disconnected'
-            };
-            break;
-        case 'paused':
-            account.stateLabel = {
-                type: 'secondary',
-                name: 'Paused'
-            };
-            break;
-        default:
-            account.stateLabel = {
-                type: 'secondary',
-                name: 'N/A'
-            };
-            break;
-    }
-
-    // Check if IMAP was disabled due to errors - override state label to show error
-    if (account.imap && account.imap.disabled && account.lastErrorState) {
-        account.stateLabel = {
-            type: 'danger',
-            name: 'Failed',
-            error: account.lastErrorState.description || account.lastErrorState.response
-        };
-    }
-
-    if (account.oauth2) {
-        account.oauth2.scopes = []
-            .concat(account.oauth2.scope || [])
-            .concat(account.oauth2.scopes || [])
-            .flatMap(entry => entry.split(/\s+/))
-            .map(entry => entry.trim())
-            .filter(entry => entry);
-
-        account.oauth2.expiresStr = account.oauth2.expires ? account.oauth2.expires.toISOString() : false;
-        account.oauth2.generatedStr = account.oauth2.generated ? account.oauth2.generated.toISOString() : false;
-
-        if (account.outlookSubscription) {
-            account.outlookSubscription.subscriptionExpiresStr = account.outlookSubscription.expirationDateTime
-                ? account.outlookSubscription.expirationDateTime.toISOString()
-                : false;
-
-            let state = account.outlookSubscription.state || {};
-
-            account.outlookSubscription.isValid =
-                state.state !== 'error' && account.outlookSubscription.expirationDateTime && account.outlookSubscription.expirationDateTime > new Date();
-
-            account.outlookSubscription.stateLabel = (state.state || '').replace(/^./, c => c.toUpperCase());
-
-            if ((state.state === 'created' && !account.outlookSubscription.expirationDateTime) || account.outlookSubscription.expirationDateTime < new Date()) {
-                account.outlookSubscription.stateLabel = 'Expired';
-            }
-        }
-    }
-
-    return account;
-}
-
-function formatServerState(state, payload) {
-    switch (state) {
-        case 'suspended':
-        case 'exited':
-        case 'disabled':
-            return {
-                type: 'warning',
-                name: state
-            };
-
-        case 'spawning':
-        case 'initializing':
-            return {
-                type: 'info',
-                name: state,
-                spinner: true
-            };
-
-        case 'listening':
-            return {
-                type: 'success',
-                name: state
-            };
-
-        case 'failed':
-            return {
-                type: 'danger',
-                name: state,
-                error: (payload && payload.error && payload.error.message) || null
-            };
-
-        default:
-            return {
-                type: 'secondary',
-                name: 'N/A'
-            };
-    }
-}
-
 async function getOpenAiError(gt) {
     let openAiError = await redis.get(`${REDIS_PREFIX}:openai:error`);
     if (openAiError) {
@@ -578,25 +347,6 @@ async function getOpenAiError(gt) {
         }
     }
     return openAiError;
-}
-
-async function getExampleDocumentsPayloads() {
-    let date = new Date().toISOString();
-
-    let examplePayloads = structuredClone(exampleDocumentsPayloads);
-
-    examplePayloads.forEach(payload => {
-        if (payload && payload.content) {
-            if (typeof payload.content.date === 'string') {
-                payload.content.date = date;
-            }
-
-            if (typeof payload.content.created === 'string') {
-                payload.content.created = date;
-            }
-        }
-    });
-    return examplePayloads;
 }
 
 async function getMailboxListing(accountObject) {
@@ -689,35 +439,15 @@ async function listPublicInterfaces(selectedAddresses) {
     });
 }
 
-async function getServerStatus(type) {
-    let serverStatus = await redis.hgetall(`${REDIS_PREFIX}${type}`);
-    let state = (serverStatus && serverStatus.state) || 'disabled';
-    let payload;
-    try {
-        payload = (serverStatus && typeof serverStatus.payload === 'string' && JSON.parse(serverStatus.payload)) || {};
-    } catch (err) {
-        // ignore
-    }
-
-    return { state, payload, label: formatServerState(state, payload) };
-}
-
 function applyRoutes(server, call) {
     // Initialize admin entity routes (webhooks, templates, gateways, tokens)
     adminEntitiesRoutes({ server, call });
 
-    // Export routes for session-authenticated UI
+    // SMTP deliverability test tool routes
+    smtpTestRoutes({ server, call });
 
-    function throwAsBoom(err) {
-        if (Boom.isBoom(err)) {
-            throw err;
-        }
-        let error = Boom.boomify(err, { statusCode: err.statusCode || 500 });
-        if (err.code) {
-            error.output.payload.code = err.code;
-        }
-        throw error;
-    }
+    // Public subscription-management (unsubscribe) routes
+    unsubscribeRoutes({ server, call });
 
     // Render-context booleans for the gmailService authMethod tab selector.
     // When `locked` is set the selector is shown but cannot be switched - the
@@ -8068,431 +7798,6 @@ Token: ${JSON.stringify(request.params.token)}`
                         .allow(false),
 
                     timezone: Joi.string().empty('').allow(false).max(255)
-                })
-            }
-        }
-    });
-
-    server.route({
-        method: 'POST',
-        path: '/admin/smtp/create-test',
-        async handler(request) {
-            let account = request.payload.account;
-
-            try {
-                request.logger.info({ msg: 'Request SMTP test', account });
-
-                let accountObject = new Account({ redis, account, call, secret: await getSecret() });
-
-                let accountData;
-                try {
-                    accountData = await accountObject.loadAccountData();
-                } catch (err) {
-                    return {
-                        error: err.message
-                    };
-                }
-
-                let headers = {
-                    'Content-Type': 'application/json',
-                    'User-Agent': `${packageData.name}/${packageData.version} (+${packageData.homepage})`
-                };
-
-                let res = await fetchCmd(`${SMTP_TEST_HOST}/test-address`, {
-                    method: 'post',
-                    body: JSON.stringify({
-                        version: packageData.version,
-                        requestor: '@postalsys/emailengine-app'
-                    }),
-                    headers,
-                    dispatcher: httpAgent.retry
-                });
-
-                if (!res.ok) {
-                    let err = new Error(res.statusText || `Invalid response: ${res.status} ${res.statusText}`);
-                    err.statusCode = res.status;
-
-                    try {
-                        err.response = await res.json();
-                    } catch (err) {
-                        // ignore
-                    }
-
-                    throw err;
-                }
-
-                let testAccount = await res.json();
-                if (!testAccount || !testAccount.user) {
-                    let err = new Error(`Invalid test account`);
-                    err.status = 500;
-
-                    try {
-                        err.response = testAccount;
-                    } catch (err) {
-                        // ignore
-                    }
-
-                    throw err;
-                }
-
-                try {
-                    let now = new Date().toISOString();
-                    let queueResponse = await accountObject.queueMessage(
-                        {
-                            account: accountData.account,
-                            subject: `Delivery test ${now}`,
-                            text: `Hello
-
-This is an automated email to test deliverability settings. If you see this email, you can safely delete it.
-
-${now}`,
-                            html: `<p>Hello</p>
-<p>This is an automated email to test deliverability settings. If you see this email, you can safely delete it.</p>
-<p>${now}</p>`,
-
-                            from: {
-                                name: accountData.name,
-                                address: accountData.email
-                            },
-                            to: [{ name: 'Delivery Test Server', address: testAccount.address }],
-                            copy: false,
-                            gateway: request.payload.gateway,
-                            feedbackKey: `${REDIS_PREFIX}test-send:${testAccount.user}`,
-                            deliveryAttempts: 1
-                        },
-                        { source: 'test' }
-                    );
-
-                    return Object.assign(testAccount, queueResponse || {});
-                } catch (err) {
-                    return {
-                        error: err.message
-                    };
-                }
-            } catch (err) {
-                request.logger.error({ msg: 'Failed to request test account', err, account });
-                return { success: false, error: err.message };
-            }
-        },
-        options: {
-            validate: {
-                options: {
-                    stripUnknown: true,
-                    abortEarly: false,
-                    convert: true
-                },
-
-                failAction,
-
-                payload: Joi.object({
-                    account: accountIdSchema.required(),
-                    gateway: Joi.string().empty('').max(256).example('sendgun').description('Gateway ID')
-                })
-            }
-        }
-    });
-
-    server.route({
-        method: 'POST',
-        path: '/admin/smtp/check-test',
-        async handler(request) {
-            let user = request.payload.user;
-
-            try {
-                request.logger.info({ msg: 'Request SMTP test response', user });
-
-                let deliveryStatus = (await redis.hgetall(`${REDIS_PREFIX}test-send:${user}`)) || {};
-                if (deliveryStatus.success === 'false') {
-                    let err = new Error(`Failed to deliver email: ${deliveryStatus.error}`);
-                    err.status = 500;
-                    throw err;
-                }
-
-                let headers = {
-                    'Content-Type': 'application/json',
-                    'User-Agent': `${packageData.name}/${packageData.version} (+${packageData.homepage})`
-                };
-
-                let res = await fetchCmd(`${SMTP_TEST_HOST}/test-address/${user}`, {
-                    method: 'get',
-                    headers,
-                    dispatcher: httpAgent.retry
-                });
-
-                if (!res.ok) {
-                    let err = new Error(res.statusText || `Invalid response: ${res.status} ${res.statusText}`);
-                    err.statusCode = res.status;
-
-                    try {
-                        err.response = await res.json();
-                    } catch (err) {
-                        // ignore
-                    }
-
-                    throw err;
-                }
-
-                let testResponse = await res.json();
-
-                if (testResponse) {
-                    let mainSig =
-                        testResponse.dkim &&
-                        testResponse.dkim.results &&
-                        testResponse.dkim.results.find(entry => entry && entry.status && entry.status.result === 'pass' && entry.status.aligned);
-
-                    if (!mainSig) {
-                        mainSig =
-                            testResponse.dkim &&
-                            testResponse.dkim.results &&
-                            testResponse.dkim.results.find(entry => entry && entry.status && entry.status.result === 'pass');
-                    }
-
-                    if (!mainSig) {
-                        mainSig = testResponse.dkim && testResponse.dkim.results && testResponse.dkim.results[0];
-                    }
-
-                    testResponse.mainSig = mainSig || {
-                        status: {
-                            result: 'none'
-                        }
-                    };
-
-                    if (testResponse.spf && testResponse.spf.status && testResponse.spf.status.comment) {
-                        testResponse.spf.status.comment = testResponse.spf.status.comment.replace(/^[^:\s]+:s*/, '');
-                    }
-                }
-
-                return testResponse;
-            } catch (err) {
-                request.logger.error({ msg: 'Failed to request test response', err, user });
-                return { status: 'error', error: err.message };
-            }
-        },
-        options: {
-            validate: {
-                options: {
-                    stripUnknown: true,
-                    abortEarly: false,
-                    convert: true
-                },
-
-                failAction,
-
-                payload: Joi.object({
-                    user: Joi.string().guid().description('Test ID')
-                })
-            }
-        }
-    });
-
-    server.route({
-        method: 'GET',
-        path: '/unsubscribe',
-        async handler(request, h) {
-            let data = Buffer.from(request.query.data, 'base64url').toString();
-            // do not check signature, validate fields in the submit step
-
-            data = JSON.parse(data);
-
-            if (!data || typeof data !== 'object' || data.act !== 'unsub') {
-                throw new Error(request.app.gt.gettext('Invalid input'));
-            }
-
-            // throws if account does not exist
-            let accountObject = new Account({ redis, account: data.acc });
-            await accountObject.loadAccountData();
-
-            return h.view(
-                'unsubscribe',
-                {
-                    pageTitleFull: request.app.gt.gettext('Subscription Management'),
-
-                    unsubscribed: await redis.hexists(`${REDIS_PREFIX}lists:unsub:entries:${data.list}`, data.rcpt),
-                    values: {
-                        listId: data.list,
-                        account: data.acc,
-                        messageId: data.msg,
-                        email: data.rcpt
-                    }
-                },
-                {
-                    layout: 'public'
-                }
-            );
-        },
-        options: {
-            auth: false,
-
-            validate: {
-                options: {
-                    stripUnknown: true,
-                    abortEarly: false,
-                    convert: true
-                },
-
-                async failAction(request, h, err) {
-                    request.logger.error({ msg: 'Failed to validate request arguments', err });
-                    let error = Boom.boomify(new Error(request.app.gt.gettext('Invalid request. Check your input and try again.')), { statusCode: 400 });
-                    if (err.code) {
-                        error.output.payload.code = err.code;
-                    }
-                    throw error;
-                },
-
-                query: Joi.object({
-                    data: Joi.string().base64({ paddingRequired: false, urlSafe: true }).required(),
-                    sig: Joi.string().base64({ paddingRequired: false, urlSafe: true })
-                })
-            }
-        }
-    });
-
-    server.route({
-        method: 'POST',
-        path: '/unsubscribe/address',
-        async handler(request, h) {
-            try {
-                // throws if account does not exist
-                let accountObject = new Account({ redis, account: request.payload.account });
-                await accountObject.loadAccountData();
-
-                let reSubscribed = false;
-
-                switch (request.payload.action) {
-                    case 'unsubscribe': {
-                        let isNew = await redis.eeListAdd(
-                            `${REDIS_PREFIX}lists:unsub:lists`,
-                            `${REDIS_PREFIX}lists:unsub:entries:${request.payload.listId}`,
-                            request.payload.listId,
-                            request.payload.email.toLowerCase().trim(),
-                            JSON.stringify({
-                                recipient: request.payload.email,
-                                account: request.payload.account,
-                                source: 'form',
-                                reason: 'unsubscribe',
-                                messageId: request.payload.messageId,
-                                remoteAddress: request.info.remoteAddress,
-                                userAgent: request.headers['user-agent'],
-                                created: new Date().toISOString()
-                            })
-                        );
-
-                        if (isNew) {
-                            await call({
-                                cmd: 'unsubscribe',
-                                account: request.payload.account,
-                                payload: {
-                                    recipient: request.payload.email,
-                                    messageId: request.payload.messageId,
-                                    listId: request.payload.listId,
-                                    remoteAddress: request.info.remoteAddress,
-                                    userAgent: request.headers['user-agent']
-                                }
-                            });
-                        }
-                        break;
-                    }
-
-                    case 'subscribe': {
-                        let removed = await redis.eeListRemove(
-                            `${REDIS_PREFIX}lists:unsub:lists`,
-                            `${REDIS_PREFIX}lists:unsub:entries:${request.payload.listId}`,
-                            request.payload.listId,
-                            request.payload.email.toLowerCase().trim()
-                        );
-
-                        if (removed) {
-                            await call({
-                                cmd: 'subscribe',
-                                account: request.payload.account,
-                                payload: {
-                                    recipient: request.payload.email,
-                                    messageId: request.payload.messageId,
-                                    listId: request.payload.listId,
-                                    remoteAddress: request.info.remoteAddress,
-                                    userAgent: request.headers['user-agent']
-                                }
-                            });
-                        }
-
-                        reSubscribed = true;
-                        break;
-                    }
-                }
-
-                return h.view(
-                    'unsubscribe',
-                    {
-                        pageTitleFull: request.app.gt.gettext('Subscription Management'),
-
-                        unsubscribed: await redis.hexists(`${REDIS_PREFIX}lists:unsub:entries:${request.payload.listId}`, request.payload.email),
-                        values: request.payload,
-                        reSubscribed
-                    },
-                    {
-                        layout: 'public'
-                    }
-                );
-            } catch (err) {
-                await request.flash({ type: 'danger', message: request.app.gt.gettext(`Couldn't process request. Try again.`) });
-                request.logger.error({ msg: 'Failed to process subscription request', err });
-
-                return h.view(
-                    'unsubscribe',
-                    {
-                        pageTitleFull: request.app.gt.gettext('Subscription Management'),
-                        unsubscribed: await redis.hexists(`${REDIS_PREFIX}lists:unsub:entries:${request.payload.listId}`, request.payload.email)
-                    },
-                    {
-                        layout: 'public'
-                    }
-                );
-            }
-        },
-        options: {
-            auth: false,
-            validate: {
-                options: {
-                    stripUnknown: true,
-                    abortEarly: false,
-                    convert: true
-                },
-
-                async failAction(request, h, err) {
-                    let errors = {};
-
-                    if (err.details) {
-                        err.details.forEach(detail => {
-                            if (!errors[detail.path]) {
-                                errors[detail.path] = detail.message;
-                            }
-                        });
-                    }
-
-                    await request.flash({ type: 'danger', message: request.app.gt.gettext(`Couldn't process request. Try again.`) });
-                    request.logger.error({ msg: 'Failed to process subscription request', err });
-
-                    return h
-                        .view(
-                            'unsubscribe',
-                            {
-                                pageTitleFull: request.app.gt.gettext('Subscription Management'),
-                                unsubscribed: await redis.hexists(`${REDIS_PREFIX}lists:unsub:entries:${request.payload.listId}`, request.payload.email),
-                                errors
-                            },
-                            {
-                                layout: 'public'
-                            }
-                        )
-                        .takeover();
-                },
-
-                payload: Joi.object({
-                    action: Joi.string().valid('subscribe', 'unsubscribe').required(),
-                    account: accountIdSchema.required(),
-                    listId: Joi.string().hostname().empty('').example('test-list').label('List ID').required(),
-                    email: Joi.string().email().empty('').required().description('Email address').required(),
-                    messageId: Joi.string().empty('').max(996).example('<test123@example.com>').description('Message ID')
                 })
             }
         }
